requirements.txt
livekit-agents[silero]>=1.3
livekit-plugins-openai>=1.3
livekit-api>=1.0
aiohttp>=3.9
